Clarify use of Summary
If the items column is not named `Item`, then you get an error when plotting the result. The error is caused by the use of `reshape2::melt(..., id.vars = "Item")` which expects the column to be named Item. This could be solved by editing the Likert code to be more adaptive but I don't know enough about R to know if that's doable. I'm assuming it was done like that for a reason.